+++ app/src/main/java/de/tolunla/ghostotp/model/Account.kt
@@ -3,7 +3,7 @@ package de.tolunla.ghostotp.model
 import de.tolunla.ghostotp.db.entity.AccountEntity
 import de.tolunla.ghostotp.db.entity.AccountEntity.Type
 
-abstract class Account(val id: Int?, val name: String, val issuer: String, val type: Type) {
+abstract class Account(val id: Long?, val name: String, val issuer: String, val type: Type) {
 
     fun getLabel(): String {
         if (issuer.isEmpty()) return name
